+++ webrtc/lib/webrtc/variant_tracker.ex
@@ -12,7 +12,7 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantTracker do
 
   @type t :: %__MODULE__{
           variant: String.t(),
-          status: :active | :inactive,
+          status: :active | :inactive | :muted,
           samples: non_neg_integer(),
           activity_cycles: non_neg_integer(),
           inactivity_cycles: non_neg_integer(),
@@ -82,6 +82,28 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantTracker do
       }
 
       maybe_active(tracker)
+    end
+  end
+
+  @spec set_variant_muted(t()) :: {:ok, t()} | {:status_changed, t(), :muted}
+  def set_variant_muted(tracker) do
+    if tracker.status == :active do
+      Membrane.Logger.debug("Variant #{inspect(tracker.variant)} is muted.")
+      tracker = %__MODULE__{tracker | status: :muted, inactivity_cycles: 0}
+      {:status_changed, tracker, :muted}
+    else
+      {:ok, tracker}
+    end
+  end
+
+  @spec set_variant_unmuted(t()) :: {:ok, t()} | {:status_changed, t(), :active}
+  def set_variant_unmuted(tracker) do
+    if tracker.status == :muted do
+      Membrane.Logger.debug("Variant #{inspect(tracker.variant)} is unmuted.")
+      tracker = %__MODULE__{tracker | status: :active, activity_cycles: 0}
+      {:status_changed, tracker, :active}
+    else
+      {:ok, tracker}
     end
   end