+++ webrtc/test/webrtc/variant_selector_test.exs
@@ -13,35 +13,17 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
     selector = create_selector()
 
     assert {selector, {:request, :medium, :variant_inactive}} =
-             VariantSelector.variant_inactive(selector, :high)
+             VariantSelector.variant_paused(selector, :high, :inactive)
 
     assert {_selector, {:request, :low, :variant_inactive}} =
-             VariantSelector.variant_inactive(selector, :medium)
-  end
-
-  # Skip explanation:
-  # After introducing automatic variant switching, this feature seems to
-  # be a out of date. This behavior will be either achieved
-  # or overwritten by the desire to always send the highest possible or target
-  # variant, that fits the bandwidth allocation
-  @tag :skip
-  test "VariantSelector selects variant being used before it became inactive" do
-    selector = create_selector()
-
-    assert {selector, {:request, :medium, :variant_inactive}} =
-             VariantSelector.variant_inactive(selector, :high)
-
-    selector = VariantSelector.set_current_variant(selector, :medium)
-
-    assert {_selector, {:request, :high, :other}} =
-             VariantSelector.variant_active(selector, :high)
+             VariantSelector.variant_paused(selector, :medium, :inactive)
   end
 
   test "target_variant/2 sets target variant and chooses it when it is active" do
     selector = create_selector()
 
     assert {selector, {:request, :medium, :variant_inactive}} =
-             VariantSelector.variant_inactive(selector, :high)
+             VariantSelector.variant_paused(selector, :high, :inactive)
 
     assert {selector, :noop} = VariantSelector.set_target_variant(selector, :medium)
 
@@ -49,23 +31,59 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
 
     # assert that variant selector doesn't request the new variant
     # even though it is better
-    assert {_selector, :noop} = VariantSelector.variant_active(selector, :high)
+    assert {_selector, :noop} = VariantSelector.variant_resumed(selector, :high)
   end
 
   test "target_variant/2 sets target variant but doesn't switch to it when it is inactive" do
     selector = create_selector()
 
-    assert {selector, :noop} = VariantSelector.variant_inactive(selector, :low)
+    assert {selector, :noop} = VariantSelector.variant_paused(selector, :low, :inactive)
     assert {selector, :noop} = VariantSelector.set_target_variant(selector, :low)
-    assert {_selector, {:request, :low, :other}} = VariantSelector.variant_active(selector, :low)
+
+    assert {_selector, {:request, :low, :variant_resumed}} =
+             VariantSelector.variant_resumed(selector, :low)
   end
 
-  @tag :skip
   test "VariantSelector doesn't select a new variant when not currently used variant is marked as inactive" do
     selector = create_selector()
 
-    assert {selector, :noop} = VariantSelector.variant_inactive(selector, :medium)
-    assert {_selector, :noop} = VariantSelector.variant_inactive(selector, :low)
+    assert {selector, :noop} = VariantSelector.variant_paused(selector, :medium, :inactive)
+    assert {_selector, :noop} = VariantSelector.variant_paused(selector, :low, :inactive)
+  end
+
+  describe "Muting Track" do
+    test "VariantSelector marks as muted the last active variant" do
+      selector = create_selector()
+      assert selector.muted_variant == :no_variant
+
+      selector = mute_all_variants(selector)
+      assert selector.muted_variant == :high
+      assert Enum.empty?(selector.active_variants)
+    end
+
+    test "VariantSelector selects another variant if the muted variant isn't active yet" do
+      selector =
+        create_selector()
+        |> mute_all_variants()
+
+      {selector, {:request, :medium, :automatic_selection}} =
+        VariantSelector.variant_resumed(selector, :medium)
+
+      assert selector.muted_variant == :high
+      assert selector.active_variants == MapSet.new([:medium])
+    end
+
+    test "VariantSelector selects muted variant once it becomes active again" do
+      selector =
+        create_selector()
+        |> mute_all_variants()
+
+      {selector, {:request, :high, :variant_unmuted}} =
+        VariantSelector.variant_resumed(selector, :high)
+
+      assert selector.muted_variant == :no_variant
+      assert selector.active_variants == MapSet.new([:high])
+    end
   end
 
   describe "Automatic variant selection in VariantSelector" do
@@ -82,31 +100,38 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
     end
 
     test "accepts any variant before receiving first allocation", %{selector: selector} do
-      Enum.each([:low, :medium, :high], fn variant ->
-        assert {_selector, {:request, ^variant, :other}} =
-                 VariantSelector.variant_active(selector, variant)
+      Enum.each([:low, :medium], fn variant ->
+        assert {_selector, {:request, ^variant, :automatic_selection}} =
+                 VariantSelector.variant_resumed(selector, variant)
       end)
+
+      assert {_selector, {:request, :high, :variant_resumed}} =
+               VariantSelector.variant_resumed(selector, :high)
     end
 
     test "frees unused allocation", %{selector: selector} do
-      assert {_selector, {:request, :low, :other}} =
-               VariantSelector.variant_active(selector, :low)
+      assert {_selector, {:request, :low, :automatic_selection}} =
+               VariantSelector.variant_resumed(selector, :low)
 
       assert_allocation_requested(:low)
     end
 
     test "requests allocation for higher variant", %{selector: selector} do
-      assert {selector, {:request, :low, :other}} = VariantSelector.variant_active(selector, :low)
+      assert {selector, {:request, :low, :automatic_selection}} =
+               VariantSelector.variant_resumed(selector, :low)
+
       assert_allocation_requested(:low)
-      assert {_selector, :noop} = VariantSelector.variant_active(selector, :medium)
+      assert {_selector, :noop} = VariantSelector.variant_resumed(selector, :medium)
       assert_allocation_requested(:medium)
     end
 
     test "doesn't select a higher variant right after lowering its allocation, before receiving confirmation from the Allocator",
          %{selector: selector} do
-      assert {selector, {:request, :low, :other}} = VariantSelector.variant_active(selector, :low)
+      assert {selector, {:request, :low, :automatic_selection}} =
+               VariantSelector.variant_resumed(selector, :low)
+
       assert_allocation_requested(:low)
-      assert {_selector, :noop} = VariantSelector.variant_active(selector, :medium)
+      assert {_selector, :noop} = VariantSelector.variant_resumed(selector, :medium)
     end
 
     test "selects the highest possible variant", %{selector: selector} do
@@ -115,7 +140,7 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
       # activate all variants
       selector =
         Enum.reduce(variants, selector, fn variant, selector ->
-          assert {selector, _action} = VariantSelector.variant_active(selector, variant)
+          assert {selector, _action} = VariantSelector.variant_resumed(selector, variant)
           selector
         end)
 
@@ -128,7 +153,7 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
         allocation = @variant_bitrates[variant] * 1.1
         assert_allocation_requested(variant)
 
-        assert {selector, {:request, ^variant, :other}} =
+        assert {selector, {:request, ^variant, :set_bandwidth_allocation}} =
                  VariantSelector.set_bandwidth_allocation(selector, allocation)
 
         selector
@@ -136,7 +161,9 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
     end
 
     test "requests to stop the track when allocation is forcefully lowered", %{selector: selector} do
-      assert {selector, {:request, :low, :other}} = VariantSelector.variant_active(selector, :low)
+      assert {selector, {:request, :low, :automatic_selection}} =
+               VariantSelector.variant_resumed(selector, :low)
+
       assert {_selector, :stop} = VariantSelector.set_bandwidth_allocation(selector, 0)
     end
 
@@ -148,18 +175,22 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
     end
 
     test "refuses to lower the allocation when on low", %{selector: selector} do
-      assert {selector, {:request, :low, :other}} = VariantSelector.variant_active(selector, :low)
+      assert {selector, {:request, :low, :automatic_selection}} =
+               VariantSelector.variant_resumed(selector, :low)
+
       assert {_selector, :noop} = VariantSelector.decrease_allocation(selector)
       assert_receive {_pid, {:decrease_allocation_request, :reject}}
     end
 
     test "lowers the allocation when on medium and it is requested", %{selector: selector} do
-      assert {selector, {:request, :low, :other}} = VariantSelector.variant_active(selector, :low)
+      assert {selector, {:request, :low, :automatic_selection}} =
+               VariantSelector.variant_resumed(selector, :low)
+
       assert_allocation_requested(:low)
-      assert {selector, :noop} = VariantSelector.variant_active(selector, :medium)
+      assert {selector, :noop} = VariantSelector.variant_resumed(selector, :medium)
       assert_allocation_requested(:medium)
 
-      assert {selector, {:request, :medium, :other}} =
+      assert {selector, {:request, :medium, :set_bandwidth_allocation}} =
                VariantSelector.set_bandwidth_allocation(
                  selector,
                  @variant_bitrates[:medium] * 1.1
@@ -198,17 +229,25 @@ defmodule Membrane.RTC.Engine.Endpoint.WebRTC.VariantSelectorTest do
 
     selector = %{selector | current_allocation: 9_999_999_999_999}
 
-    assert {selector, {:request, :high, :other}} = VariantSelector.variant_active(selector, :high)
+    assert {selector, {:request, :high, :variant_resumed}} =
+             VariantSelector.variant_resumed(selector, :high)
 
     selector = VariantSelector.set_current_variant(selector, :high)
 
-    assert {selector, :noop} = VariantSelector.variant_active(selector, :medium)
-    assert {selector, :noop} = VariantSelector.variant_active(selector, :low)
+    assert {selector, :noop} = VariantSelector.variant_resumed(selector, :medium)
+    assert {selector, :noop} = VariantSelector.variant_resumed(selector, :low)
     selector
   end
 
   defp assert_allocation_requested(variant) do
     bitrate = @variant_bitrates[variant] * 1.1
     assert_receive {:request_allocation, _pid, ^bitrate}
+  end
+
+  defp mute_all_variants(selector) do
+    Enum.reduce([:low, :medium, :high], selector, fn variant, selector ->
+      {selector, :noop} = VariantSelector.variant_paused(selector, variant, :muted)
+      selector
+    end)
   end
 end