+++ engine/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.23.0-dev
 * Add RTCP sender reports [#393](https://github.com/fishjam-dev/membrane_rtc_engine/pull/393)
+* Add reason to `TrackVariantPaused` event [#392](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/392)
 
 ## 0.22.0
 * Update deps [#374](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/374)